requirements.local.txt
# Core packages needed to run the application locally
# Install with: pip install -r requirements.local.txt

# Already installed (skip these):
# fastapi==0.116.1
# uvicorn==0.35.0
# pydantic==2.11.7
# pydantic-settings==2.11.0
# sqlalchemy==2.0.44

# Additional packages to install:
python-dotenv>=1.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.11.1
aiosqlite>=0.19.0

# HTTP client
httpx>=0.24.1

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.9.0
tika>=2.6.0
pytesseract>=0.3.10
python-magic>=0.4.27

# AI/ML - PyTorch (CPU version)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

# AI/ML - NLP
transformers>=4.30.0
spacy>=3.6.0
nltk>=3.8.1
scikit-learn>=1.3.0
sentence-transformers>=2.2.2

# Optional (for full functionality)
elasticsearch>=8.9.0
elasticsearch-dsl>=8.9.0
redis>=4.6.0
celery>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.1

# Monitoring
prometheus-client>=0.17.1
